# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Hotel Guest Check-in API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List all bookings
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, completed, all]
          required: false
        - in: query
          name: search
          schema:
            type: string
          required: false
        - in: query
          name: page
          schema:
            type: integer
          required: false
        - in: query
          name: limit
          schema:
            type: integer
          required: false
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingList"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a new booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingInput"
      responses:
        "201":
          description: Created booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
...
  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get a single booking with guests
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingDetail"
        "404":
          description: Not found
    patch:
      operationId: updateBooking
      tags: [bookings]
      summary: Update booking details
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingUpdate"
      responses:
        "200":
          description: Updated booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "404":
          description: Not found
    delete:
      operationId: deleteBooking
      tags: [bookings]
      summary: Delete a booking
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /bookings/{id}/checkin-link:
    get:
      operationId: getCheckinLink
      tags: [bookings]
      summary: Get check-in link for a booking
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Check-in link info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckinLink"
        "404":
          description: Not found
...
  /bookings/{id}/guests:
    get:
      operationId: listBookingGuests
      tags: [guests]
      summary: List all guests for a booking
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Guest records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GuestRecord"
  /bookings/{id}/export:
    get:
      operationId: getExportData
      tags: [bookings]
      summary: Export booking guest details (client handles PDF/Excel rendering)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Export data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExportData"
...
  /bookings/bulk-status:
    post:
      operationId: bulkUpdateBookingStatus
      tags: [bookings]
      summary: Bulk update booking status
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingUpdateStatus"
      responses:
        "200":
          description: Updated count
          content:
            application/json:
              schema:
                type: object
                required: [updated]
                properties:
                  updated:
                    type: integer
  /bookings/bulk-delete:
    post:
      operationId: bulkDeleteBookings
      tags: [bookings]
      summary: Bulk delete bookings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ids]
              properties:
                ids:
                  type: array
                  items:
                    type: integer
      responses:
        "200":
          description: Deleted count
          content:
            application/json:
              schema:
                type: object
                required: [deleted]
                properties:
                  deleted:
                    type: integer
  /bookings/{id}/duplicate:
    post:
      operationId: duplicateBooking
      tags: [bookings]
      summary: Duplicate a booking
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Duplicated booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "404":
          description: Not found
components:
  schemas:
...
    Booking:
      type: object
      required: [id, bookingRef, guestName, roomNumber, checkInDate, checkOutDate, numberOfGuests, status, checkinToken, createdAt]
      properties:
        id:
          type: integer
        bookingRef:
          type: string
        guestName:
          type: string
        phone:
          type: ["string", "null"]
        roomNumber:
          type: string
        checkInDate:
          type: string
        checkOutDate:
          type: string
        numberOfGuests:
          type: integer
        status:
          type: string
          enum: [pending, completed]
        checkinToken:
          type: string
        notes:
          type: ["string", "null"]
        guestsSubmitted:
          type: integer
        createdAt:
          type: string
    BookingDetail:
      type: object
      required: [id, bookingRef, guestName, roomNumber, checkInDate, checkOutDate, numberOfGuests, status, checkinToken, createdAt, guests]
      properties:
        id:
          type: integer
        bookingRef:
          type: string
        guestName:
          type: string
        phone:
          type: ["string", "null"]
        roomNumber:
          type: string
        checkInDate:
          type: string
        checkOutDate:
          type: string
        numberOfGuests:
          type: integer
        status:
          type: string
          enum: [pending, completed]
        checkinToken:
          type: string
        notes:
          type: ["string", "null"]
        guestsSubmitted:
          type: integer
        createdAt:
          type: string
        guests:
          type: array
          items:
            $ref: "#/components/schemas/GuestRecord"
    BookingList:
      type: object
      required: [bookings, total]
      properties:
        bookings:
          type: array
          items:
            $ref: "#/components/schemas/Booking"
        total:
          type: integer
    BookingInput:
      type: object
      required: [guestName, roomNumber, checkInDate, checkOutDate, numberOfGuests]
      properties:
        guestName:
          type: string
          minLength: 1
        phone:
          type: string
        roomNumber:
          type: string
          minLength: 1
        checkInDate:
          type: string
        checkOutDate:
          type: string
        numberOfGuests:
          type: integer
          minimum: 1
        notes:
          type: string
    BookingUpdate:
      type: object
      properties:
        guestName:
          type: string
        phone:
          type: string
        roomNumber:
          type: string
        checkInDate:
          type: string
        checkOutDate:
          type: string
        numberOfGuests:
          type: integer
        status:
          type: string
          enum: [pending, completed]
        notes:
          type: string
    CheckinLink:
      type: object
      required: [token, url, whatsappUrl, smsUrl]
      properties:
        token:
          type: string
        url:
          type: string
        whatsappUrl:
          type: string
        smsUrl:
          type: string
...
    GuestRecord:
      type: object
      required: [id, bookingId, fullName, phone, email, address, nationality, idType, isForeignNational, createdAt]
      properties:
        id:
          type: integer
        bookingId:
          type: integer
        fullName:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        nationality:
          type: string
        idType:
          type: string
          enum: [aadhaar, passport, driving_licence, voter_id, other]
        idNumber:
          type: ["string", "null"]
        idImageUrl:
          type: ["string", "null"]
        signatureUrl:
          type: ["string", "null"]
        isForeignNational:
          type: boolean
        numberOfGuests:
          type: ["integer", "null"]
        visaNumber:
          type: ["string", "null"]
        visaType:
          type: ["string", "null"]
        visaExpiry:
          type: ["string", "null"]
        portOfEntry:
          type: ["string", "null"]
        arrivalDate:
          type: ["string", "null"]
        passportNumber:
          type: ["string", "null"]
        passportExpiry:
          type: ["string", "null"]
        createdAt:
          type: string
...
    ExportData:
      type: object
      required: [booking, guests, generatedAt]
      properties:
        booking:
          $ref: "#/components/schemas/Booking"
        guests:
          type: array
          items:
            $ref: "#/components/schemas/GuestRecord"
        generatedAt:
          type: string
...
    BookingUpdateStatus:
      type: object
      required: [ids, status]
      properties:
        ids:
          type: array
          items:
            type: integer
        status:
          type: string
          enum: [pending, completed]