# lib/api-zod/src/generated/types/bookingUpdateStatus.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Hotel Guest Check-in API
 * OpenAPI spec version: 0.1.0
 */

export type BookingUpdateStatus = typeof BookingUpdateStatus[keyof typeof BookingUpdateStatus];


export const BookingUpdateStatus = {
  pending: 'pending',
  completed: 'completed',
} as const;