# lib/api-zod/src/generated/types/checkinLink.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Hotel Guest Check-in API
 * OpenAPI spec version: 0.1.0
 */

export interface CheckinLink {
  token: string;
  url: string;
  whatsappUrl: string;
  smsUrl: string;
}